refactor(taskflow): extract feature.settings
Move SetTheme/SetLatency/SetFailureRate/SetBotEnabled/SetOnline from
action/Actions.kt, appSettingsReducer from reducer/RootReducers.kt,
SettingsScreen and SettingsSlider from ui/ into the new
feature.settings package. Update all consumer imports tree-wide.
