+++ examples/taskflow/composeApp/src/commonMain/kotlin/org/reduxkotlin/sample/taskflow/reducer/RootReducers.kt
@@ -8,17 +8,11 @@ import org.reduxkotlin.sample.taskflow.action.LoadAccountsSucceeded
 import org.reduxkotlin.sample.taskflow.action.LoginFailed
 import org.reduxkotlin.sample.taskflow.action.LoginRequested
 import org.reduxkotlin.sample.taskflow.action.LogoutAccount
-import org.reduxkotlin.sample.taskflow.action.SetBotEnabled
-import org.reduxkotlin.sample.taskflow.action.SetFailureRate
-import org.reduxkotlin.sample.taskflow.action.SetLatency
-import org.reduxkotlin.sample.taskflow.action.SetOnline
-import org.reduxkotlin.sample.taskflow.action.SetTheme
 import org.reduxkotlin.sample.taskflow.action.StartLogin
 import org.reduxkotlin.sample.taskflow.action.SwitchAccount
 import org.reduxkotlin.sample.taskflow.core.AccountId
 import org.reduxkotlin.sample.taskflow.core.AccountSummary
 import org.reduxkotlin.sample.taskflow.core.Action
-import org.reduxkotlin.sample.taskflow.core.AppSettingsModel
 import org.reduxkotlin.sample.taskflow.model.AccountsModel
 import org.reduxkotlin.sample.taskflow.model.AuthFlowModel
 
@@ -76,31 +70,6 @@ public fun accountsReducer(model: AccountsModel, action: Action): AccountsModel 
             )
         }
     }
-
-    else -> model
-}
-
-/**
- * Pure root-store reducer for the [AppSettingsModel] slice (theme + fake-service config).
- *
- * Returns the same [model] instance unchanged for actions it does not handle.
- *
- * @param model the current app-settings slice.
- * @param action the dispatched action.
- * @return the next app-settings slice, or [model] unchanged when [action] is not handled.
- */
-public fun appSettingsReducer(model: AppSettingsModel, action: Action): AppSettingsModel = when (action) {
-    is SetTheme -> model.copy(theme = action.theme)
-
-    is SetLatency -> model.copy(
-        fakeService = model.fakeService.copy(latencyMinMs = action.minMs, latencyMaxMs = action.maxMs),
-    )
-
-    is SetFailureRate -> model.copy(fakeService = model.fakeService.copy(failureRate = action.rate))
-
-    is SetBotEnabled -> model.copy(fakeService = model.fakeService.copy(botEnabled = action.enabled))
-
-    is SetOnline -> model.copy(fakeService = model.fakeService.copy(online = action.online))
 
     else -> model
 }
