+++ examples/taskflow/composeApp/src/commonMain/kotlin/org/reduxkotlin/sample/taskflow/store/AppStore.kt
@@ -15,19 +15,19 @@ import org.reduxkotlin.sample.taskflow.action.LoadAccountsSucceeded
 import org.reduxkotlin.sample.taskflow.action.LoginFailed
 import org.reduxkotlin.sample.taskflow.action.LoginRequested
 import org.reduxkotlin.sample.taskflow.action.LogoutAccount
-import org.reduxkotlin.sample.taskflow.action.SetBotEnabled
-import org.reduxkotlin.sample.taskflow.action.SetFailureRate
-import org.reduxkotlin.sample.taskflow.action.SetLatency
-import org.reduxkotlin.sample.taskflow.action.SetOnline
-import org.reduxkotlin.sample.taskflow.action.SetTheme
 import org.reduxkotlin.sample.taskflow.action.StartLogin
 import org.reduxkotlin.sample.taskflow.action.SwitchAccount
 import org.reduxkotlin.sample.taskflow.core.AppSettingsModel
+import org.reduxkotlin.sample.taskflow.feature.settings.SetBotEnabled
+import org.reduxkotlin.sample.taskflow.feature.settings.SetFailureRate
+import org.reduxkotlin.sample.taskflow.feature.settings.SetLatency
+import org.reduxkotlin.sample.taskflow.feature.settings.SetOnline
+import org.reduxkotlin.sample.taskflow.feature.settings.SetTheme
+import org.reduxkotlin.sample.taskflow.feature.settings.appSettingsReducer
 import org.reduxkotlin.sample.taskflow.infra.platform.mainNotificationContext
 import org.reduxkotlin.sample.taskflow.model.AccountsModel
 import org.reduxkotlin.sample.taskflow.model.AuthFlowModel
 import org.reduxkotlin.sample.taskflow.reducer.accountsReducer
-import org.reduxkotlin.sample.taskflow.reducer.appSettingsReducer
 import org.reduxkotlin.sample.taskflow.reducer.authFlowReducer
 
 /**
